[v2] Extract field filtering and make generic
